+++ src/packed_n_seq.rs
@@ -8,7 +8,7 @@ pub struct PackedNSeq<'s> {
     pub ambiguous: BitSeq<'s>,
 }
 
-#[derive(Clone, Debug, MemSize, MemDbg)]
+#[derive(Clone, Debug, MemSize, MemDbg, Default)]
 #[cfg_attr(feature = "pyo3", pyo3::pyclass)]
 #[cfg_attr(feature = "epserde", derive(epserde::Epserde))]
 pub struct PackedNSeqVec {
@@ -28,6 +28,11 @@ impl<'s> PackedNSeq<'s> {
 
 /// Implement a subset of `SeqVec` for `PackedNSeqVec`.
 impl PackedNSeqVec {
+    pub fn clear(&mut self) {
+        self.seq.clear();
+        self.ambiguous.clear();
+    }
+
     pub fn random(len: usize, n_frac: f32) -> Self {
         let seq = PackedSeqVec::random(len);
         let ambiguous = BitSeqVec::random(len, n_frac);
@@ -39,6 +44,11 @@ impl PackedNSeqVec {
             seq: self.seq.as_slice(),
             ambiguous: self.ambiguous.as_slice(),
         }
+    }
+
+    pub fn push_ascii(&mut self, seq: &[u8]) {
+        self.seq.push_ascii(seq);
+        self.ambiguous.push_ascii(seq);
     }
 
     pub fn from_ascii(seq: &[u8]) -> Self {